+++ src/contexts/AuthContext.tsx
@@ -1,28 +1,8 @@
-import React, { createContext } from 'react'
+import React from 'react'
 import type { ReactNode } from 'react'
 import { useAuth } from '../hooks/useAuth'
 import { authUtils } from '../utils/auth'
-import type { AuthUser } from '../utils/auth'
-import type { User } from 'firebase/auth'
-
-// Define the shape of our auth context
-export interface AuthContextType {
-  // Auth state
-  user: AuthUser | null
-  loading: boolean
-  error: string | null
-  isAuthenticated: boolean
-  isGuest: boolean
-  
-  // Auth actions
-  signIn: (email: string, password: string) => Promise<{ user: User | null; error: Error | null }>
-  signUp: (email: string, password: string, displayName?: string) => Promise<{ user: User | null; error: Error | null }>
-  signOut: () => Promise<{ error: Error | null }>
-  resetPassword: (email: string) => Promise<{ error: Error | null }>
-}
-
-// Create the context
-export const AuthContext = createContext<AuthContextType | undefined>(undefined)
+import { AuthContext, type AuthContextType } from './AuthContextData'
 
 // AuthProvider component that wraps your app
 interface AuthProviderProps {
